Added ButtonsTrait for Facebook Message and also Card. Fixed typo's Reformatted slightly for StyleCI
