+++ src/Components/Card.php
@@ -3,12 +3,13 @@
 namespace NotificationChannels\Facebook\Components;
 
 use NotificationChannels\Facebook\Exceptions\CouldNotCreateCard;
-use NotificationChannels\Facebook\Exceptions\CouldNotCreateMessage;
-use NotificationChannels\Facebook\Exceptions\CouldNotSendNotification;
+use NotificationChannels\Facebook\Traits\ButtonsTrait;
 
 
 class Card implements \JsonSerializable
 {
+    use ButtonsTrait;
+
     /** @var string Card Title */
     protected $title;
 
@@ -21,11 +22,8 @@ class Card implements \JsonSerializable
     /** @var string Subtitle */
     protected $subtitle;
 
-    /** @var array Call to Action Buttons */
-    public $buttons = [];
-
     /**
-     * Create a Card
+     * Create a Card.
      *
      * @return static
      */
@@ -35,7 +33,7 @@ class Card implements \JsonSerializable
     }
 
     /**
-     * Create Card constructor
+     * Create Card constructor.
      */
     public function __construct($title = '')
     {
@@ -55,28 +53,12 @@ class Card implements \JsonSerializable
             throw CouldNotCreateCard::titleLimitExceeded($this->title);
         }
         $this->title = $title;
+
         return $this;
     }
 
     /**
-     * Add up to 3 call to action buttons.
-     *
-     * @param array $buttons
-     *
-     * @return $this
-     * @throws CouldNotSendNotification
-     */
-    public function buttons(array $buttons = [])
-    {
-        if (count($buttons) > 3) {
-            throw CouldNotCreateMessage::messageButtonsLimitExceeded();
-        }
-        $this->buttons = $buttons;
-        return $this;
-    }
-
-    /**
-     * Set Card Item Url
+     * Set Card Item Url.
      *
      * @param $item_url
      * @return $this
@@ -84,11 +66,12 @@ class Card implements \JsonSerializable
     public function url($item_url)
     {
         $this->item_url = $item_url;
+
         return $this;
     }
 
     /**
-     * Set Card Image Url
+     * Set Card Image Url.
      *
      * @param $image_url
      * Image ration should be 1.91:1
@@ -97,11 +80,12 @@ class Card implements \JsonSerializable
     public function image($image_url)
     {
         $this->image_url = $image_url;
+
         return $this;
     }
 
     /**
-     * Set Card Subtitle
+     * Set Card Subtitle.
      *
      * @param $subtitle
      * @throws CouldNotCreateCard
@@ -109,15 +93,17 @@ class Card implements \JsonSerializable
      */
     public function subtitle($subtitle)
     {
-        if (mb_strlen($subtitle) > 80)
+        if (mb_strlen($subtitle) > 80) {
             throw CouldNotCreateCard::subtitleLimitExceeded($this->title);
+        }
         $this->subtitle = $subtitle;
+
         return $this;
     }
 
 
     /**
-     * Builds payload and returns an array
+     * Builds payload and returns an array.
      *
      * @return array
      * @throws CouldNotCreateCard
@@ -126,21 +112,26 @@ class Card implements \JsonSerializable
     {
         $payload = [];
 
-        if (!isset($this->title))
+        if (!isset($this->title)) {
             throw CouldNotCreateCard::titleNotProvided();
+        }
         $payload['title'] = $this->title;
 
-        if (isset($this->item_url))
+        if (isset($this->item_url)) {
             $payload['item_url'] = $this->item_url;
+        }
 
-        if (isset($this->image_url))
+        if (isset($this->image_url)) {
             $payload['image_url'] = $this->image_url;
+        }
 
-        if (isset($this->subtitle))
+        if (isset($this->subtitle)) {
             $payload['subtitle'] = $this->subtitle;
+        }
 
-        if (count($this->buttons) > 0)
+        if (count($this->buttons) > 0) {
             $payload['buttons'] = $this->buttons;
+        }
 
         return $payload;
     }
@@ -154,4 +145,5 @@ class Card implements \JsonSerializable
     {
         return $this->toArray();
     }
+
 }
