+++ src/Enums/AttachmentType.php
@@ -2,16 +2,15 @@
 
 namespace NotificationChannels\Facebook\Enums;
 
-
 /**
- * Class AttachmentType
- * @package NotificationChannels\Facebook\Enums
- * List of possible Attachment Types
+ * Class AttachmentType.
  */
 class AttachmentType
 {
+
     const IMAGE = 'image';
     const AUDIO = 'audio';
     const VIDEO = 'video';
     const FILE = 'file';
+
 }
