+++ src/Enums/ButtonType.php
@@ -2,14 +2,14 @@
 
 namespace NotificationChannels\Facebook\Enums;
 
-
 /**
- * Class ButtonType
- * @package NotificationChannels\Facebook\Enums
+ * Class ButtonType.
  */
 class ButtonType
 {
+
     const WEB_URL = 'web_url';
     const POSTBACK = 'postback';
     const PHONE_NUMBER = 'phone_number';
+
 }
