+++ src/Enums/NotificationType.php
@@ -3,13 +3,12 @@
 namespace NotificationChannels\Facebook\Enums;
 
 /**
- * Class NotificationType
- * @package NotificationChannels\Facebook\Enums
- * List of possible Notification Type values
+ * Class NotificationType.
  */
 class NotificationType
 {
     const REGULAR = 'REGULAR';
     const NO_PUSH = 'NO_PUSH';
     const SILENT_PUSH = 'SILENT_PUSH';
+
 }
