+++ src/Exceptions/CouldNotSendNotification.php
@@ -45,5 +45,4 @@ class CouldNotSendNotification extends \Exception
         return new static('The communication with Facebook failed. Reason: '.$exception->getMessage());
     }
 
-
 }
